+++ spec/project/object/native_target_spec.rb
@@ -122,6 +122,83 @@ module ProjectSpecs
           @target.build_configuration_list.set_setting('OTHER_LDFLAGS', %w(-framework UIKit))
           @target.resolved_build_setting('OTHER_LDFLAGS', true).should == { 'Release' => %w(-framework UIKit), 'Debug' => %w(-framework UIKit) }
         end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution recursively' do
+          project_xcconfig = @project.new_file(fixture_path('project.xcconfig'))
+          @project.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = project_xcconfig }
+          @target.resolved_build_setting('PRODUCT_BUNDLE_IDENTIFIER', true).should == { 'Release' => 'com.cocoapods.app', 'Debug' => 'com.cocoapods.app.dev' }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution appending' do
+          project_xcconfig = @project.new_file(fixture_path('project.xcconfig'))
+          @project.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = project_xcconfig }
+          @target.resolved_build_setting('CONFIG_APPEND', true).should == { 'Release' => 'PROJECT_XCCONFIG_VALUE_Release', 'Debug' => 'PROJECT_XCCONFIG_VALUE_Debug' }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution in same level' do
+          @target.build_configuration_list.set_setting('TARGET_USER_DEFINED', '${TARGET_USER_DEFINED_2}')
+          @target.build_configuration_list.set_setting('TARGET_USER_DEFINED_2', 'TARGET_USER_DEFINED_VALUE')
+          @target.resolved_build_setting('TARGET_USER_DEFINED', true).should == { 'Release' => 'TARGET_USER_DEFINED_VALUE', 'Debug' => 'TARGET_USER_DEFINED_VALUE' }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution' do
+          project_xcconfig = @project.new_file(fixture_path('project.xcconfig'))
+          @project.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = project_xcconfig }
+          @target.resolved_build_setting('DEVELOPMENT_TEAM', true).should == { 'Release' => 'PROJECT_XCCONFIG_VALUE', 'Debug' => 'PROJECT_XCCONFIG_VALUE' }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution: target xcconfig referencing target xcconfig' do
+          target_xcconfig = @project.new_file(fixture_path('target.xcconfig'))
+          @target.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = target_xcconfig }
+          expected_value_release = 'User Defined xcconfig target Release'
+          expected_value_debug = 'User Defined xcconfig target Debug'
+          @target.resolved_build_setting('TARGET_REFERENCE_XCCONFIG_TARGET', true).should == { 'Release' => expected_value_release, 'Debug' => expected_value_debug }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution: target xcconfig referencing target' do
+          target_xcconfig = @project.new_file(fixture_path('target.xcconfig'))
+          @target.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = target_xcconfig }
+          @target.build_configuration_list.set_setting('TARGET_USER_DEFINED', 'TARGET_USER_DEFINED_VALUE')
+          @target.resolved_build_setting('TARGET_REFERENCE_TARGET', true).should == { 'Release' => 'TARGET_USER_DEFINED_VALUE', 'Debug' => 'TARGET_USER_DEFINED_VALUE' }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution: target xcconfig referencing project xcconfig' do
+          project_xcconfig = @project.new_file(fixture_path('project.xcconfig'))
+          @project.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = project_xcconfig }
+          target_xcconfig = @project.new_file(fixture_path('target.xcconfig'))
+          @target.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = target_xcconfig }
+          expected_value_debug = 'User Defined xcconfig project Debug'
+          expected_value_release = 'User Defined xcconfig project Release'
+          @target.resolved_build_setting('TARGET_REFERENCE_XCCONFIG_PROJECT', true).should == { 'Release' => expected_value_release, 'Debug' => expected_value_debug }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution: target xcconfig referencing project' do
+          @project.build_configuration_list.set_setting('PROJECT_USER_DEFINED', 'PROJECT_USER_DEFINED_VALUE')
+          target_xcconfig = @project.new_file(fixture_path('target.xcconfig'))
+          @target.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = target_xcconfig }
+          @target.resolved_build_setting('TARGET_REFERENCE_PROJECT', true).should == { 'Release' => 'PROJECT_USER_DEFINED_VALUE', 'Debug' => 'PROJECT_USER_DEFINED_VALUE' }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution: project xcconfig referencing project xcconfig' do
+          project_xcconfig = @project.new_file(fixture_path('project.xcconfig'))
+          @project.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = project_xcconfig }
+          expected_value_release = 'User Defined xcconfig project Release'
+          expected_value_debug = 'User Defined xcconfig project Debug'
+          @target.resolved_build_setting('PROJECT_REFERENCE_XCCONFIG_PROJECT', true).should == { 'Release' => expected_value_release, 'Debug' => expected_value_debug }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution: target xcconfig referencing project' do
+          @project.build_configuration_list.set_setting('PROJECT_USER_DEFINED', 'PROJECT_USER_DEFINED_VALUE')
+          project_xcconfig = @project.new_file(fixture_path('project.xcconfig'))
+          @project.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = project_xcconfig }
+          @target.resolved_build_setting('PROJECT_REFERENCE_PROJECT', true).should == { 'Release' => 'PROJECT_USER_DEFINED_VALUE', 'Debug' => 'PROJECT_USER_DEFINED_VALUE' }
+        end
+
+        it 'returns the resolved build setting string value for a given key considering variable substitution: project referencing target' do
+          @project.build_configuration_list.set_setting('PROJECT_REFERENCE_TARGET', '$(TARGET_USER_DEFINED)')
+          @target.build_configuration_list.set_setting('TARGET_USER_DEFINED', 'TARGET_USER_DEFINED_VALUE')
+          @target.resolved_build_setting('PROJECT_REFERENCE_TARGET', true).should == { 'Release' => 'TARGET_USER_DEFINED_VALUE', 'Debug' => 'TARGET_USER_DEFINED_VALUE' }
+        end
       end
 
       #----------------------------------------#